refactor: extract metadata helper to util
