+++ packages/astro/src/core/render/util.ts
@@ -1,4 +1,7 @@
 import npath from 'path-browserify';
+import type { ModuleNode, ViteDevServer } from 'vite';
+import type { Metadata } from '../../runtime/server/metadata.js';
+import { MARKDOWN_IMPORT_FLAG } from '../../vite-plugin-markdown/index.js';
 
 /** Normalize URL to its canonical form */
 export function createCanonicalURL(url: string, base?: string): URL {
@@ -36,3 +39,47 @@ const cssRe = new RegExp(
 		.join('|')})($|\\?)`
 );
 export const isCSSRequest = (request: string): boolean => cssRe.test(request);
+
+// During prod builds, some modules have dependencies we should preload by hand
+// Ex. markdown files imported asynchronously or via Astro.glob(...)
+// This calls each md file's $$loadMetadata to discover those dependencies
+// and writes all results to the input `metadata` object
+export async function collectMdMetadata(
+	metadata: Metadata,
+	modGraph: ModuleNode,
+	viteServer: ViteDevServer
+) {
+	const importedModules = [...(modGraph?.importedModules ?? [])];
+	await Promise.all(
+		importedModules.map(async (importedModule) => {
+			const importedModGraph = importedModule.id
+				? viteServer.moduleGraph.getModuleById(importedModule.id)
+				: null;
+			// if the imported module has a graph entry, recursively collect metadata
+			if (importedModGraph) await collectMdMetadata(metadata, importedModGraph, viteServer);
+
+			if (!importedModule?.id?.endsWith(MARKDOWN_IMPORT_FLAG)) return;
+
+			const mdSSRMod = await viteServer.ssrLoadModule(importedModule.id);
+			const mdMetadata = (await mdSSRMod.$$loadMetadata?.()) as Metadata;
+			if (!mdMetadata) return;
+
+			for (let mdMod of mdMetadata.modules) {
+				mdMod.specifier = mdMetadata.resolvePath(mdMod.specifier);
+				metadata.modules.push(mdMod);
+			}
+			for (let mdHoisted of mdMetadata.hoisted) {
+				metadata.hoisted.push(mdHoisted);
+			}
+			for (let mdHydrated of mdMetadata.hydratedComponents) {
+				metadata.hydratedComponents.push(mdHydrated);
+			}
+			for (let mdClientOnly of mdMetadata.clientOnlyComponents) {
+				metadata.clientOnlyComponents.push(mdClientOnly);
+			}
+			for (let mdHydrationDirective of mdMetadata.hydrationDirectives) {
+				metadata.hydrationDirectives.add(mdHydrationDirective);
+			}
+		})
+	);
+}
